// src/models/attributes/enums/Unit.ts
export enum Unit {
  // Air & Gas units
  CFM = 'CFM',           // Cubic Feet per Minute
  SCFM = 'SCFM',         // Standard Cubic Feet per Minute
  
  // Water & Chemical units
  GPM = 'GPM',           // Gallons per Minute
  LPM = 'LPM',           // Liters per Minute
  
  // Electricity units
  Amps = 'Amps',         // Amperes
  Volts = 'Volts',       // Volts
  
  // Signal units
  Hertz = 'Hertz',       // Frequency (Hz)
  Decibels = 'Decibels', // Signal strength (dB)
}